InferenceDAO design added.
