+++ src/main/java/com/courseApp/reviewService/RecommendationRequestProcessor.java
@@ -1,6 +1,8 @@
 package com.courseApp.reviewService;
 
 import com.courseApp.constants.Constants;
+import com.courseApp.dao.InferenceDAO;
+import com.courseApp.dao.InferenceDaoImpl;
 import org.apache.http.Header;
 import org.apache.http.HttpEntity;
 import org.apache.http.client.methods.CloseableHttpResponse;
@@ -59,70 +61,12 @@ public class RecommendationRequestProcessor implements UseRecommendationUpdate, 
     @Override
     public double modelInference(String reviewString) throws Exception {
 
-        String res = queryResult(sentRequest(reviewString));
-        return Double.parseDouble(res);
+        return new InferenceDaoImpl().modelInference(reviewString);
     }
 
 
-    /**
-     * Sent post request to our ML inference server.
-     *
-     * @param reviewString string of user review
-     * @return String of inference query url
-     * @throws IOException exception
-     */
-    private String sentRequest(String reviewString) throws IOException {
-        CloseableHttpClient httpclient = HttpClients.createDefault();
-        HttpPost httpPost = new HttpPost(PREDICTION_URL);
-        String JSON_STRING= JSON_BODY.replaceFirst(DOC_IDENTIFIER, reviewString);
-        httpPost.setHeader(SUBSCRIPTION_KEY, SUBSCRIPTION_VALUE);
-        httpPost.setHeader(CONTENT_TYPE, JSON_TYPE);
-        // Add authentication ket to the request header
-        HttpEntity stringEntity = new StringEntity(JSON_STRING, ContentType.APPLICATION_JSON);
-        httpPost.setEntity(stringEntity);
-
-        try (CloseableHttpResponse response2 = httpclient.execute(httpPost)) {
-            Header[] headers = response2.getHeaders(OPERATION_LOCATION);
-            return headers[0].getValue();
-        }
-    }
 
 
-    /**
-     * Get the inference result from the inference server.
-     *
-     * @param queryLocation url for getting the inference result
-     * @return String of classification result
-     * @throws Exception Server exception
-     */
-    private String queryResult(String queryLocation) throws Exception {
-        CloseableHttpClient httpclient = HttpClients.createDefault();
-        HttpGet httpGet = new HttpGet(queryLocation);
-        httpGet.setHeader(SUBSCRIPTION_KEY, SUBSCRIPTION_VALUE);
-        httpGet.setHeader(CONTENT_TYPE, JSON_TYPE);
-        // Add authentication ket to the request header
-
-        int count = 0;
-        int maxTries = 10; // try ten times
-        final Pattern pattern = Pattern.compile(REGEX_CLASSIFICATION, Pattern.CASE_INSENSITIVE); //use regex to match the result
 
 
-        while(true){
-            try (CloseableHttpResponse response2 = httpclient.execute(httpGet)) {
-                HttpEntity httpEntity = response2.getEntity();
-                String s = EntityUtils.toString(httpEntity);
-                final Matcher matcher = pattern.matcher(s);
-                if (matcher.find())
-                {
-                    return matcher.group(0).substring(matcher.group(0).length() - 1); //get last char, which is the classification
-                }
-                else{
-                    TimeUnit.SECONDS.sleep(2);
-                    if (++count == maxTries){ // time out
-                        throw new Exception(INFERENCE_SERVER_ERROR);
-                    }
-                }
-            }
-        }
-    }
 }
